Fixed a bug in the Instagram service when there is not a caption.
Not all instagram pictures have captions on them, so we must check that there is a caption before trying to access it or we get PHP warnings that bugger up the JSON parsing.

Also fixed the output so that when there is a caption "user: caption" is displayed under the picture, but when there is not then just "user" is displayed (no colon).