Fixed incorrect pagination.
Previously it was incorrectly assumed that the last id in the list would
be the `max_id` for the next query. This works for recent posts, however
for the `tags/xxx/media/recent` query this is not neccesarily true.

Instead, all Instagram endpoints return a pagination object with this
information. This change uses that information in setting `max_id` and
`min_id`.

Tags endpoint: http://instagram.com/developer/endpoints/tags/#
Pagination info: http://instagram.com/developer/endpoints/#